:recycle: refactor: move programmatic API from bin/generate-changelog to lib/changelog
